api/requirements.txt
anyio==3.6.2 ; python_version >= "3.10" and python_version < "4.0"
cachetools==5.2.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.9.24 ; python_version >= "3.10" and python_version < "4"
charset-normalizer==2.1.1 ; python_version >= "3.10" and python_version < "4"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
fastapi==0.87.0 ; python_version >= "3.10" and python_version < "4.0"
google-api-core==2.10.2 ; python_version >= "3.10" and python_version < "4.0"
google-api-core[grpc]==2.10.2 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.14.1 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-appengine-logging==1.1.6 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-audit-log==0.2.4 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-core==2.3.2 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-firestore==2.7.2 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-logging==3.2.5 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.57.0 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos[grpc]==1.57.0 ; python_version >= "3.10" and python_version < "4.0"
grpc-google-iam-v1==0.12.4 ; python_version >= "3.10" and python_version < "4.0"
grpcio-status==1.50.0 ; python_version >= "3.10" and python_version < "4.0"
grpcio==1.50.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
proto-plus==1.22.1 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.21.9 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.2.8 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.2 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.1 ; python_version >= "3.10" and python_version < "4"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.13 ; python_version >= "3.10" and python_version < "4"
uvicorn==0.20.0 ; python_version >= "3.10" and python_version < "4.0"
